Hide HTML anchor link on h1 headings
This commit hides the default anchor link provided by the theme on h1 headings. Since we only use one h1 heading per page, it makes little sense to offer a link for both our pages as well the h1 of our pages—because it boils down to the exact same thing, basically. Anchor links remain active on h2, h3, and so on…