Update Gemfile
This commit switches the github-pages gem off and turns the jekyll one on, in order for our site to run the latest version of Jekyll available. Since we'll use a GitHub action to build our site, we won't be limited to Jekyll 3 anymore—as we were when we were using GitHub Pages to build our site.