Update _config.yml
This commit adds the jekyll-remote-theme plugin to the list of plugins we're using. Since we won't be relying on GitHub Pages to build our site, we have to install the jekyll-remote-theme plugin manually (and list it in our config file too). That plugin will allow us to use the Just the docs theme, although this theme isn't natively supported by GitHub Pages—we can't use it as a gem-based theme, but we can use it as a remote-theme.